Improve error handling capibilities
These changes improve the way we detected scraper errors. Previously, we
only handled errors that were set by us explicitly, while completely
ignoring errors that happened during the execution of a spider. That
resulted in marking scrapers that failed with "[W08] No new files",
which is false.

This commit introduces a Scrapy extension that handles spider failures
by setting a corresponding stat entry with exception's name and message.
Later, during stat parsing, we're converting that entry to "[E00]
<exception>: <message>" error.

We're also using the "log_count/ERROR" stat to fail with a message
"[E00] Internal error" in case we don't have any information on why the
spider failed.